Rename permissions middleware file.
